+++ src/page_components/__init__.py
@@ -1,4 +1,4 @@
-__version__ = "0.1"
+__version__ = "1.0.0"
 
 
 from .page_component import *  # noqa