+++ src/page_components/page_component.py
@@ -1,12 +1,10 @@
 import abc
-from typing import *
+from typing import Any, Dict
 
 import asset_definitions
 import django.core.exceptions
 import django.forms
 import django.template.loader
-import django.utils.encoding
-import django.utils.six
 
 TemplateContext = Dict[str, Any]
 
@@ -17,13 +15,9 @@ __all__ = (
 )
 
 
-@django.utils.encoding.python_2_unicode_compatible
-class PageComponent(
-    django.utils.six.with_metaclass(abc.ABCMeta, asset_definitions.MediaDefiningClass)
-):
+class PageComponent(abc.ABC, asset_definitions.MediaDefiningClass):
     @abc.abstractmethod
-    def render(self):
-        # type: () -> str
+    def render(self) -> str:
         return ""
 
     def __str__(self):
@@ -34,8 +28,7 @@ class TemplatePageComponent(PageComponent):
 
     template_name = None
 
-    def render(self):
-        # type: () -> str
+    def render(self) -> str:
         template_name = self.get_template_name()
         context_data = self.get_context_data()
         return django.template.loader.render_to_string(
@@ -43,8 +36,7 @@ class TemplatePageComponent(PageComponent):
         )
 
     # noinspection PyMethodMayBeStatic
-    def get_template_name(self):
-        # type: () -> str
+    def get_template_name(self) -> str:
         if not self.template_name:
             raise django.core.exceptions.ImproperlyConfigured(
                 "template_name is missing"
@@ -52,7 +44,6 @@ class TemplatePageComponent(PageComponent):
         return self.template_name
 
     # noinspection PyMethodMayBeStatic
-    def get_context_data(self, **kwargs):
-        # type: () -> TemplateContext
+    def get_context_data(self, **kwargs) -> TemplateContext:
         context_data = dict(kwargs)
         return context_data