+++ CHANGES.rst
@@ -2,7 +2,14 @@
 Changes
 =======
 
+1.0.0
+=====
+
+- Add Django 3.x and 4.x compatibility
+- Drop support for Python 2
+- Python 3.6 is the minimum requirement
+
 0.1
 ===
 
- - Initial release
+- Initial release